add message protocol and sqlite
Application model could send messages to message center which dispatch
messages.
Messages could be send to local data center and app model back and
forth sync or async.
So message model will be totally  independent as a bridge between app
and data.
The framework and easily be modified to support network if the message
model send message to socket or http model. App model doesn’t have to
worry about where the message finally goes to because it only
communicate with message center.